requirements.txt
openai>=1.58.1
python-dotenv>=1.0.1
python-pptx>=1.0.2
requests>=2.32.3
streamlit>=1.41.1